+++ src/RcppExports.cpp
@@ -12876,14 +12876,14 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// ladder_clamp_counts_differentiated_tf24
-std::vector<std::vector<double>> ladder_clamp_counts_differentiated_tf24(plant::RcppR6::RcppR6<plant::SCM<plant::TF24_Strategy<double>, plant::TF24_Environment<double> > > obj_);
-RcppExport SEXP _plant_ladder_clamp_counts_differentiated_tf24(SEXP obj_SEXP) {
+// census_clamp_counts_differentiated_tf24
+std::vector<std::vector<double>> census_clamp_counts_differentiated_tf24(plant::RcppR6::RcppR6<plant::SCM<plant::TF24_Strategy<double>, plant::TF24_Environment<double> > > obj_);
+RcppExport SEXP _plant_census_clamp_counts_differentiated_tf24(SEXP obj_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< plant::RcppR6::RcppR6<plant::SCM<plant::TF24_Strategy<double>, plant::TF24_Environment<double> > > >::type obj_(obj_SEXP);
-    rcpp_result_gen = Rcpp::wrap(ladder_clamp_counts_differentiated_tf24(obj_));
+    rcpp_result_gen = Rcpp::wrap(census_clamp_counts_differentiated_tf24(obj_));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -12897,14 +12897,14 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// ladder_curvature_margin_tf24
-std::vector<double> ladder_curvature_margin_tf24(plant::RcppR6::RcppR6<plant::SCM<plant::TF24_Strategy<double>, plant::TF24_Environment<double> > > obj_);
-RcppExport SEXP _plant_ladder_curvature_margin_tf24(SEXP obj_SEXP) {
+// census_curvature_margin_tf24
+std::vector<double> census_curvature_margin_tf24(plant::RcppR6::RcppR6<plant::SCM<plant::TF24_Strategy<double>, plant::TF24_Environment<double> > > obj_);
+RcppExport SEXP _plant_census_curvature_margin_tf24(SEXP obj_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< plant::RcppR6::RcppR6<plant::SCM<plant::TF24_Strategy<double>, plant::TF24_Environment<double> > > >::type obj_(obj_SEXP);
-    rcpp_result_gen = Rcpp::wrap(ladder_curvature_margin_tf24(obj_));
+    rcpp_result_gen = Rcpp::wrap(census_curvature_margin_tf24(obj_));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -14301,9 +14301,9 @@ static const R_CallMethodDef CallEntries[] = {
     {"_plant_census_clear_diagnostics_tf24", (DL_FUNC) &_plant_census_clear_diagnostics_tf24, 1},
     {"_plant_census_operating_point_names_tf24", (DL_FUNC) &_plant_census_operating_point_names_tf24, 0},
     {"_plant_census_clamp_counts_tf24", (DL_FUNC) &_plant_census_clamp_counts_tf24, 1},
-    {"_plant_ladder_clamp_counts_differentiated_tf24", (DL_FUNC) &_plant_ladder_clamp_counts_differentiated_tf24, 1},
+    {"_plant_census_clamp_counts_differentiated_tf24", (DL_FUNC) &_plant_census_clamp_counts_differentiated_tf24, 1},
     {"_plant_census_clamp_names_tf24", (DL_FUNC) &_plant_census_clamp_names_tf24, 0},
-    {"_plant_ladder_curvature_margin_tf24", (DL_FUNC) &_plant_ladder_curvature_margin_tf24, 1},
+    {"_plant_census_curvature_margin_tf24", (DL_FUNC) &_plant_census_curvature_margin_tf24, 1},
     {"_plant_FF16_oderunner_individual_internals", (DL_FUNC) &_plant_FF16_oderunner_individual_internals, 1},
     {"_plant_K93_oderunner_individual_internals", (DL_FUNC) &_plant_K93_oderunner_individual_internals, 1},
     {"_plant_TF24_oderunner_individual_internals", (DL_FUNC) &_plant_TF24_oderunner_individual_internals, 1},